LayoutTests/editing/pasteboard/emacs-ctrl-a-k-y-expected.txt

Tests:
ctrl-a, ctrl-k, and ctrl-y
Expected Results:

one two three four five
one two three four five
